refactor run_scene_with_inspector
